Specify ref in build workflow
So… when the scrape workflow triggers the build workflow, the build
workflow doesn't check out the latest commit. This appears to be an
issue because it inherits the SHA from the calling workflow.

Since the scrape workflow creates a new comimt, the existing ref is the
next-to-last commit. Some folks figured out a workaround:

https://github.com/actions/checkout/issues/439#issuecomment-1464126822