Account for spaces in URLs
Specific example, from 2023-06-12:

```md
[Tianyao Yang](mailto: yangtian1994 at gmail dot com)
```

The markdown-it parser won't recognize this URL as valid, so it doesn't
generate `<a href="…">…</a>` markup. We get around this by replacing the
space characters with URL-encoded `%20`.

In this specific case, it's still not an email address, but markdown-it
does its job and the link will open a user's mail client.